gkelog: add an opentelemetry TraceSpanExtractor
Similar to Opencensus, this one is its own submodule as well.
It locks the current head version of opentelemetry because they renamed
the context-extraction method and I don't want to rename it later.

I had to remove the "sampled" variant of the test because there's
currently no way that I could find to enable sampling without doing
unholy things to the otel SpanContext struct (mostly setting flags
manually).

It looks like they have some missing plumbing, that should get resolved,
so there's a TODO waiting for us when that's working.